Add support for the real-time events API
Zulip's real-time events API [1] allows registering to a stream of
events that occur on the server in real-time. The most useful of these
events are "message" events, that are fired when a new message is seen
by the user.

This PR adds support for the real-time events API in OZulip through the
new [Events] module, whose interface is thoroughly documented in the PR.

The most convenient function in the [Events] module is [interact], which
evaluates a Lwt thread in response to messages from trusted senders that
mention the current user, and sends the resulting of evaluating the
thread as a reply to the original message.

[1]: https://zulip.com/api/real-time-events